// frontend/tdk-app/src/app/model/ExcelApplication.ts
export class ExcelApplication {
    status: string = '';
    tdkTitle: string = '';
    tdkTitle_EN: string = '';
    facultySectionName: string = '';
    facultyName: string = '';
    equipments: string = '';
    summary: string = '';
    language: string = '';
    /*Supervisor*/
    supervisorName: string = '';
    supervisorPosition: string = '';
    supervisorFaculty: string = '';
    supervisorInstitute: string = '';
    /*Author*/
    authorName: string = '';
    authorFacultyName: string = '';
    authorSpecialization: string = '';
    authorYear: string = '';
    authorNeptunCode: string = '';
    authorIdNumber: string = '';
    authorTypeOfTheSpecialization: string = '';
    authorTaxStatus: string = '';
    authorIsEmployee: string = '';
    authorTaxIdentificationNumber: string = '';
    authorBirthPlace: string = '';
    authorBirthDate: string = '';
    authorNameOfTheMother: string = '';
    authorZipCode: string = '';
    authorTown: string = '';
    authorStreetAndNumber: string = '';
    authorTelephoneNumber: string = '';
    authorEmail: string = '';
    authorBankAccountNumber: string = '';
    /*Projects*/
    project: string = '';
    /*Reviewers*/
    reviewerName: string = '';
    reviewerEmail: string = '';
    reviewerWorkplace: string = '';

}