// frontend/tdk-app/src/app/model/Author.ts
export class Author {
    // tslint:disable-next-line: variable-name
    name: string = '';
    facultyName: string = '';
    specialization: string = '';
    year: number;
    isGraduate: string = '';
    neptunCode: string = '';
    idNumber: string = '';
    typeOfTheSpecialization: string = '';
    taxStatus: string = '';
    isEmployee: string = '';
    taxIdentificationNumber: string = '';
    birthPlace: string = '';
    birthDate: Date;
    nameOfTheMother: string = '';
    zipCode: string = '';
    town: string = '';
    streetAndNumber: string = '';
    telephoneNumber: string = '';
    email: string = '';
    bankAccountNumber: string = '';
}